requirements.txt
streamlit
langchain
selenium
beautifulsoup4
lxml
html5lib
python-dotenv
requests
google-generativeai

